validation_engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
web3==6.11.1
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
ipfshttpclient==0.8.0
python-jose==3.3.0
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
eth-account==0.9.0
eth-typing==3.5.1
eth-utils==2.3.0 